third_party/WebKit/LayoutTests/fast/dom/document-set-title-single-DOMSubtreeModified-expected.txt

On setting document.title, only one DOMSubtreeModified event is created.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.title is "new"
PASS i == 1 is true
PASS successfullyParsed is true

TEST COMPLETE
